LightOperations: Move calculate_testcase_name() into a new file
- start using a seperate file for pytest related operations
- with this we can avoid future cyclic import problem between
operations.py and testcase_parameters.py
