+++ tests/python_functional/functional_tests/conftest.py
@@ -24,8 +24,8 @@ import logging
 
 from pathlib2 import Path
 
-from src.common.operations import calculate_testcase_name
 from src.common.operations import copy_file
+from src.common.pytest_operations import calculate_testcase_name
 
 logger = logging.getLogger(__name__)
