+++ tests/python_functional/src/testcase_parameters/testcase_parameters.py
@@ -22,7 +22,7 @@
 #############################################################################
 from pathlib2 import Path
 
-from src.common.operations import calculate_testcase_name
+from src.common.pytest_operations import calculate_testcase_name
 
 
 class TestcaseParameters(object):
